--------//--------
git-svn-id: svn://svn.code.sf.net/p/eos-movrec/code/trunk@85 ab1d2490-7c84-4f7d-a8b8-f67e5ba0a9ad

+++ capturewnd.h
@@ -59,7 +59,7 @@ private:
 	QPoint MousePressPoint;
 	double ZPD_x, ZPD_y;
 	bool ZoomRectMoving;
-	double** FocusArea;
+	int** FocusArea;
 	QRect FocusAreaRect;
 	QMutex FocusMutex;
 	QString Text;
